gr: add support for securely-managed falcons
Start securely-managed falcons using the CPUCTL_ALIAS register since
CPUCTL is protected in that case.

This solution (doing secure boot in GR) is temporary. In the future the
PMU firmware will be able to reboot falcons on demand, but this requires
the PMU firmware to be released first. This solution at least allows GR
to be enabled.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>